store/wizard: extract filesystem types to submodule
Types are the foundation that utilities and selectors depend on.
Moving them first establishes the filesystem submodule boundary
and lets subsequent commits build on it cleanly.
