+++ src/Components/CreateImageWizard/steps/FileSystem/fscPolicies.ts
@@ -1,4 +1,4 @@
-import { MountpointPoliciesType } from './fscTypes';
+import { MountpointPoliciesType } from '@/store/slices/wizard';
 
 // Policies according to https://github.com/osbuild/images/blob/main/pkg/policies/policies.go
 export const MountpointPolicies: MountpointPoliciesType = {
