+++ src/Components/CreateImageWizard/utilities/parseSizeUnit.ts
@@ -1,5 +1,6 @@
+import { Units } from '@/store/slices/wizard';
+
 import { UNIT_GIB, UNIT_MIB } from '../../../constants';
-import { Units } from '../steps/FileSystem/fscTypes';
 
 export const parseSizeUnit = (bytesize: string) => {
   let size;
