+++ src/store/slices/wizard/index.ts
@@ -1,1 +1,4 @@
 export * from './slice';
+
+// submodules
+export * from './filesystem';
